change object to string
