+++ src/Agoda.IoC.Core/IKeyedComponentFactory.cs
@@ -2,8 +2,8 @@ namespace Agoda.IoC.Core
 {
     public interface IKeyedComponentFactory<out T>
     {
-        T GetByKey(object key);
-        bool IsRegistered(object key);
-        T TryGetByKey(object key);
+        T GetByKey(string key);
+        bool IsRegistered(string key);
+        T TryGetByKey(string key);
     }
 }
