+++ src/Agoda.IoC.NetCore/NetCoreKeyedComponentResolver.cs
@@ -6,19 +6,10 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace Agoda.IoC.NetCore
 {
-    public class NetCoreKeyedRegistrations<T>
-    {
-        public NetCoreKeyedRegistrations(IDictionary<object, Type> registrations)
-        {
-            Registrations = registrations;
-        }
-
-        public IDictionary<object, Type> Registrations { get; }
-    }
     public class NetCoreKeyedComponentResolver<T> : IKeyedComponentResolver<T>
     {
         private readonly IServiceProvider _container;
-        private readonly IDictionary<object, Type> _registrations;
+        private readonly IDictionary<string, Type> _registrations;
 
         public NetCoreKeyedComponentResolver(IServiceProvider container, NetCoreKeyedRegistrations<T> registrations)
         {
@@ -26,14 +17,14 @@ namespace Agoda.IoC.NetCore
             _registrations = registrations.Registrations;
         }
 
-        public T Resolve(object key)
+        public T Resolve(string key)
         {
             if (!_registrations.TryGetValue(key, out var implementationType))
                 throw new ArgumentException($"Service name '{key}' is not registered");
             return (T)_container.GetService(implementationType);
         }
 
-        public bool IsRegistered(object key)
+        public bool IsRegistered(string key)
         {
             return !_registrations.TryGetValue(key, out var implementationType);
         }
